Fix: Change short types to int types
This commit changes short java types to int java types for those fields hat are 1 byte long. In Java, there are no unsigned variables so even a short type is 2 bytes long, it always performs sign extension for those values > 127 leading into errors when parsing the 'identifier' field, for example

+++ src/main/java/org/danielmartinez/radius/packet/Attribute.java
@@ -10,7 +10,7 @@ public class Attribute {
      * Type field: It identifies the type of RADIUS attribute
      * Length: 1 Byte
      */
-    private short type;
+    private int type;
 
     /**
      * Length field: It indicates the length of this attribute including the Type, Length and Value fields.
@@ -25,7 +25,7 @@ public class Attribute {
     private byte[] value;
 
     // Constructor
-    public Attribute(short type, int length, byte[] value) {
+    public Attribute(int type, int length, byte[] value) {
         this.type = type;
         this.length = length;
         this.value = value;
@@ -36,15 +36,15 @@ public class Attribute {
     }
 
     // Getters and setters
-    public short getType() { return type; }
-    public void setType(short type) {
+    public int getType() { return type; }
+    public void setType(int type) {
         this.type = type;
     }
 
     public int getLength() {
         return length;
     }
-    public void setLength(short length) {
+    public void setLength(int length) {
         this.length = length;
     }
